[AutoComplete] Create Fragments annotation component (#297)
* Create fragments annotator

* Add margins

* Keep separate state in command and fragments tool to avoid refresh

* Rename filters files to autocomplete

* Add callback to logical form

* Update README with setup instructions

* Change output path in autocomplete preprocess script
