+++ tools/data_processing/preprocess_datasets_for_autocomplete.py
@@ -52,7 +52,7 @@ if __name__ == "__main__":
     # Default is assuming script is run from template tool backend
     parser.add_argument("--annotations_dir_path", default="../../../../craftassist/agent/datasets/full_data/")
     parser.add_argument("--commands_path", default="commands.txt")
-    parser.add_argument("--json_out_path", default="command_dict_pairs.json")
+    parser.add_argument("--json_out_path", default="../frontend/src/command_dict_pairs.json")
     args = parser.parse_args()
     print("*** Loading data pairs from {} ***".format(args.annotations_dir_path))
     all_commands = create_JSON_from_txt(args.annotations_dir_path)
